backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.32.2
pydantic>=2.9.0
python-dotenv>=1.0.0
